requirements.txt
pydub==0.25.1
PyPDF2==3.0.1
python-dotenv==1.0.1
